clean main page
