+++ package.json
@@ -16,6 +16,7 @@
         "eslint-plugin-vue": "^9.22.0",
         "laravel-vite-plugin": "^1.0.0",
         "postcss": "^8.4.35",
+        "postcss-nesting": "^12.1.0",
         "prettier": "^3.2.5",
         "tailwindcss": "^3.4.1",
         "vite": "^5.0.0"
