+++ postcss.config.js
@@ -1,6 +1,8 @@
 export default {
-  plugins: {
-    tailwindcss: {},
-    autoprefixer: {},
-  },
-}
+    plugins: {
+        "postcss-import": {},
+        "tailwindcss/nesting": "postcss-nesting",
+        tailwindcss: {},
+        autoprefixer: {},
+    },
+};
