+++ vite.config.js
@@ -17,6 +17,12 @@ export default defineConfig({
             "@": fileURLToPath(
                 new URL("./resources/js/Pages", import.meta.url),
             ),
+            "@dashboardCentralPageComponents": fileURLToPath(
+                new URL(
+                    "./resources/js/Pages/components/centralPages/dashboard",
+                    import.meta.url,
+                ),
+            ),
             "ziggy-js": path.resolve("vendor/tightenco/ziggy"),
         },
     },
